fixed an error that occurred If you didn't have any dms with the person

+++ index.py
@@ -8,11 +8,18 @@ class Purger(discord.Client):
         return message.author == self.user
     
     async def on_connect(self):
+        print('Now purging')
+        if not self.user.friends:
+            print('No friends to purge.')
+            return
         for friend in self.user.friends:
             print(f'Purging DMS with {str(friend)}')
+            if not friend.dm_channel:
+                continue
             async for message in friend.dm_channel.history(limit=100).filter(self.check):
                 await message.delete()
-    
+        print('Finished purging.')
+        
     def main(self):
         os.system('cls')
         token = input('Token: ')
